LPD-77976 Move productPurchase method to it's own RestController
