+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-etc-spring-boot/client-extension.yaml
@@ -8,7 +8,7 @@ liferay-marketplace-etc-spring-boot-analytics-provisioning:
 liferay-marketplace-etc-spring-boot-marketplace-product-purchase:
     name: Liferay Marketplace Etc Spring Boot Marketplace Product Purchase
     oAuth2ApplicationExternalReferenceCode: liferay-marketplace-etc-spring-boot-oaua
-    resourcePath: /marketplace/product/purchase
+    resourcePath: /object/action/product-purchase
     type: objectAction
 liferay-marketplace-etc-spring-boot-marketplace-product-submit:
     name: Liferay Marketplace Etc Spring Boot Marketplace Product Submit
