+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-etc-spring-boot/src/main/java/com/liferay/marketplace/MarketplaceRestController.java
@@ -14,7 +14,6 @@ import com.liferay.headless.admin.user.client.resource.v1_0.AccountResource;
 import com.liferay.headless.admin.user.client.resource.v1_0.AccountRoleResource;
 import com.liferay.headless.admin.user.client.resource.v1_0.PostalAddressResource;
 import com.liferay.headless.admin.user.client.resource.v1_0.UserAccountResource;
-import com.liferay.headless.commerce.admin.catalog.client.dto.v1_0.Catalog;
 import com.liferay.headless.commerce.admin.catalog.client.dto.v1_0.Currency;
 import com.liferay.headless.commerce.admin.catalog.client.dto.v1_0.Product;
 import com.liferay.headless.commerce.admin.catalog.client.resource.v1_0.CurrencyResource;
@@ -25,12 +24,9 @@ import com.liferay.headless.commerce.admin.order.client.pagination.Page;
 import com.liferay.headless.commerce.admin.order.client.pagination.Pagination;
 import com.liferay.headless.commerce.admin.order.client.resource.v1_0.OrderItemResource;
 import com.liferay.headless.commerce.admin.order.client.resource.v1_0.OrderResource;
-import com.liferay.marketplace.constants.MarketplaceConstants;
 import com.liferay.marketplace.model.PublisherAssetLink;
-import com.liferay.marketplace.model.SalesforceOpportunity;
 import com.liferay.marketplace.service.KoroneikiService;
 import com.liferay.marketplace.service.MarketplaceService;
-import com.liferay.marketplace.service.SalesforceService;
 import com.liferay.marketplace.util.MarketplaceUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -251,106 +247,6 @@ public class MarketplaceRestController extends BaseRestController {
 		return ResponseEntity.ok(account);
 	}
 
-	@PostMapping("product/purchase")
-	public void postProductPurchase(
-			@AuthenticationPrincipal Jwt jwt, @RequestBody String json)
-		throws Exception {
-
-		if (_log.isInfoEnabled()) {
-			_log.info("POST product purchase " + json);
-		}
-
-		JSONObject jsonObject = new JSONObject(json);
-
-		JSONObject commerceOrderJSONObject = jsonObject.getJSONObject(
-			"commerceOrder");
-
-		String paymentMethod = commerceOrderJSONObject.getString(
-			"paymentMethod");
-		int paymentStatus = commerceOrderJSONObject.getInt("paymentStatus");
-
-		Order order = _marketplaceService.getOrder(
-			commerceOrderJSONObject.getLong("id"));
-
-		if ((Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_MONEY_ORDER) &&
-			 (paymentStatus ==
-				 MarketplaceConstants.ORDER_PAYMENT_STATUS_PENDING)) ||
-			(Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_PAYPAL) &&
-			 (paymentStatus ==
-				 MarketplaceConstants.ORDER_PAYMENT_STATUS_COMPLETED))) {
-
-			_sendOrderPurchasedNotification(order);
-		}
-
-		_sendNotification(order, json);
-
-		if ((paymentStatus !=
-				MarketplaceConstants.ORDER_PAYMENT_STATUS_COMPLETED) &&
-			(paymentStatus !=
-				MarketplaceConstants.ORDER_PAYMENT_STATUS_NOT_REQUIRED)) {
-
-			if (_log.isInfoEnabled()) {
-				_log.info(
-					"Skipping POST product purchase for order " +
-						commerceOrderJSONObject.getLong("id") +
-							" because payment status is not completed");
-			}
-
-			return;
-		}
-
-		_marketplaceService.updateOrder(
-			null, order.getId(), MarketplaceConstants.ORDER_STATUS_PROCESSING);
-
-		String orderTypeExternalReferenceCode =
-			order.getOrderTypeExternalReferenceCode();
-		Map<String, String> productSpecificationsMap =
-			_marketplaceService.getProductSpecificationsMap(
-				_marketplaceService.getOrderProductId(order));
-
-		if (Objects.equals(orderTypeExternalReferenceCode, "ADDONS")) {
-			_setUpAddOns(jwt, order, productSpecificationsMap);
-
-			_marketplaceService.updateOrder(
-				null, order.getId(),
-				MarketplaceConstants.ORDER_STATUS_COMPLETED);
-		}
-
-		if (Objects.equals(orderTypeExternalReferenceCode, "CLOUD_APP") ||
-			Objects.equals(orderTypeExternalReferenceCode, "COMPOSITE_APP") ||
-			Objects.equals(
-				orderTypeExternalReferenceCode, "LOW_CODE_CONFIGURATION") ||
-			Objects.equals(orderTypeExternalReferenceCode, "OTHER")) {
-
-			_marketplaceService.updateOrder(
-				null, order.getId(),
-				MarketplaceConstants.ORDER_STATUS_COMPLETED);
-		}
-
-		if (Objects.equals(
-				orderTypeExternalReferenceCode, "CLIENT_EXTENSION") ||
-			Objects.equals(
-				order.getOrderTypeExternalReferenceCode(), "DXP_APP")) {
-
-			if (Objects.equals(
-					productSpecificationsMap.get("price-model"), "Free")) {
-
-				_marketplaceService.updateOrder(
-					null, order.getId(),
-					MarketplaceConstants.ORDER_STATUS_COMPLETED);
-
-				return;
-			}
-
-			_setUpProductEntitlements(
-				jwt, productSpecificationsMap.get("license-type"), order);
-		}
-	}
-
 	@PostMapping("product/submit")
 	public void postProductSubmit(
 			@AuthenticationPrincipal Jwt jwt, @RequestBody String json)
@@ -566,24 +462,6 @@ public class MarketplaceRestController extends BaseRestController {
 		return accountRole.getId();
 	}
 
-	private String _getExchangeRate(Order order) {
-		Map<String, String> customFields =
-			(Map<String, String>)order.getCustomFields();
-
-		JSONObject orderMetadataJSONObject = new JSONObject(
-			customFields.getOrDefault("order-metadata", "{}"));
-
-		if (!Objects.equals(order.getCurrencyCode(), "USD") ||
-			!orderMetadataJSONObject.has("exchangeRate")) {
-
-			return "Not applicable";
-		}
-
-		double exchangeRate = orderMetadataJSONObject.getDouble("exchangeRate");
-
-		return "1 USD = " + String.format("%.5f", exchangeRate) + " EUR";
-	}
-
 	private File _getPublisherAssetFile(String publisherAssetURL)
 		throws Exception {
 
@@ -681,341 +559,6 @@ public class MarketplaceRestController extends BaseRestController {
 		}
 	}
 
-	private void _sendNotification(Order order, String json) throws Exception {
-		String paymentMethod = order.getPaymentMethod();
-		int paymentStatus = order.getPaymentStatus();
-
-		if ((Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_MONEY_ORDER) &&
-			 (paymentStatus ==
-				 MarketplaceConstants.ORDER_PAYMENT_STATUS_PENDING)) ||
-			(Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_PAYPAL) &&
-			 (paymentStatus ==
-				 MarketplaceConstants.ORDER_PAYMENT_STATUS_COMPLETED))) {
-
-			_sendOrderPurchasedNotification(order);
-		}
-
-		String orderEmailDescription = "";
-		String orderTypeExternalReferenceCode =
-			order.getOrderTypeExternalReferenceCode();
-		Map<String, String> productSpecificationsMap =
-			_marketplaceService.getProductSpecificationsMap(
-				_marketplaceService.getOrderProductId(order));
-
-		if (Objects.equals(orderTypeExternalReferenceCode, "CDP")) {
-			orderEmailDescription =
-				"<p>Your workspace is being created now!</p>" +
-					"<p>Click the button below to go to your dashboard and check the status of your environment. " +
-						"You can start using it as soon as it is ready.</p>";
-		}
-		else if (Objects.equals(orderTypeExternalReferenceCode, "DXP")) {
-			String priceModel = productSpecificationsMap.get("price-model");
-
-			if (Objects.equals(priceModel, "Free")) {
-				orderEmailDescription =
-					"<p>Your app is ready for download.</p>" +
-						"<p>To find your app download, find your Order ID and choose Manage, " +
-							"then Download LPKG.</p>";
-			}
-			else if (Objects.equals(priceModel, "Paid")) {
-				orderEmailDescription =
-					"<p>Your app is ready for download.</p>" +
-						"<p>To access your download, find your Order ID and select Manage, " +
-							"then <b>Download LPKG.</b> " +
-								"Please note that a <b>valid license is also required to activate </b> " +
-									"and use the application." + "</p>";
-			}
-		}
-
-		if (Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_PAYPAL) &&
-			(paymentStatus ==
-				MarketplaceConstants.ORDER_PAYMENT_STATUS_COMPLETED)) {
-
-			String orderButtonText = "Go to Dashboard";
-
-			_sendOrderConfirmationNotification(
-				order, orderEmailDescription, orderButtonText, json);
-		}
-
-		if (Objects.equals(
-				paymentMethod,
-				MarketplaceConstants.ORDER_PAYMENT_METHOD_MONEY_ORDER) &&
-			(paymentStatus ==
-				MarketplaceConstants.ORDER_PAYMENT_STATUS_COMPLETED)) {
-
-			String paymentEmailDescription =
-				"<p>You are all set 🚀 You <b>can start using all the premium features </b> " +
-					"of your Customer Data Platform right away. " +
-						"Click the button below to access your CDP and enjoy the full experience." +
-							"</p>";
-
-			String paymentButtonText = "Launch LDP";
-
-			_sendPaymentApprovedNotification(
-				order, paymentEmailDescription, paymentButtonText, json);
-		}
-	}
-
-	private void _sendOrderConfirmationNotification(
-			Order order, String emailDescription, String buttonText,
-			String json)
-		throws Exception {
-
-		OrderItem[] orderItems = order.getOrderItems();
-
-		OrderItem orderItem = orderItems[0];
-
-		if (orderItem == null) {
-			return;
-		}
-
-		if (_log.isInfoEnabled()) {
-			_log.info(
-				"Sending order confirmation notification for order " +
-					order.getId());
-		}
-
-		Product product = _marketplaceService.getProductBySkuId(
-			orderItem.getSkuId());
-
-		Catalog catalog = _marketplaceService.getCatalog(
-			product.getCatalogId());
-
-		JSONObject jsonObject = new JSONObject(json);
-
-		JSONObject modelCPDefinitionJSONObject = jsonObject.getJSONObject(
-			"modelCPDefinition");
-
-		_marketplaceService.postNotificationQueueEntry(
-			order.getCreatorEmailAddress(), "MARKETPLACE-ORDER-CONFIRMATION",
-			new HashMapBuilder<String, String>().put(
-				"[%APP_NAME%]",
-				product.getName(
-				).get(
-					"en_US"
-				)
-			).put(
-				"[%CATALOG_NAME%]", catalog.getName()
-			).put(
-				"[%COMMERCEORDER_AUTHOR_EMAIL_ADDRESS%]",
-				order.getCreatorEmailAddress()
-			).put(
-				"[%ORDER_ID%]", String.valueOf(order.getId())
-			).put(
-				"[%PRODUCT_THUMBNAIL%]",
-				new URL(
-					StringBundler.concat(
-						lxcDXPServerProtocol, "://", lxcDXPMainDomain,
-						product.getThumbnail())
-				).toString(
-				).replaceAll(
-					"(?<=accounts/)-?\\d+(?=/images)", "-1"
-				)
-			).put(
-				"[%EMAIL_DESCRIPTION%]", emailDescription
-			).put(
-				"[%BUTTON_TEXT%]", buttonText
-			).put(
-				"[%APP_PRICE%]", order.getTotalFormatted()
-			).put(
-				"[%CPDEFINITION_ID%]",
-				String.valueOf(
-					modelCPDefinitionJSONObject.getLong("CPDefinitionId"))
-			).build());
-	}
-
-	private void _sendOrderPurchasedNotification(Order order) throws Exception {
-		OrderItem[] orderItems = order.getOrderItems();
-
-		OrderItem orderItem = orderItems[0];
-
-		if (orderItem == null) {
-			return;
-		}
-
-		if (_log.isInfoEnabled()) {
-			_log.info(
-				"Sending purchased order notification for order " +
-					order.getId());
-		}
-
-		com.liferay.headless.commerce.admin.order.client.dto.v1_0.Account
-			account = order.getAccount();
-
-		BillingAddress billingAddress = _marketplaceService.getBillingAddress(
-			order.getId());
-
-		Product product = _marketplaceService.getProductBySkuId(
-			orderItem.getSkuId());
-
-		Catalog catalog = _marketplaceService.getCatalog(
-			product.getCatalogId());
-
-		Map<String, String> productSpecificationsMap =
-			_marketplaceService.getProductSpecificationsMap(
-				product.getProductId());
-
-		_marketplaceService.postNotificationQueueEntry(
-			null, "MARKETPLACE-INVOICE-ORDER-SUBMIT-TEMPLATE",
-			new HashMapBuilder<String, String>().put(
-				"[%ACCOUNT_ID%]", String.valueOf(account.getId())
-			).put(
-				"[%ACCOUNT_NAME%]", account.getName()
-			).put(
-				"[%APP_NAME%]",
-				product.getName(
-				).get(
-					"en_US"
-				)
-			).put(
-				"[%APP_TYPE%]",
-				productSpecificationsMap.get(
-					"type"
-				).replace(
-					"-", " "
-				)
-			).put(
-				"[%BILLING_ADDRESS_FORMATTED%]",
-				String.join(
-					", ", billingAddress.getStreet1(), billingAddress.getCity(),
-					billingAddress.getRegionISOCode(),
-					billingAddress.getCountryISOCode())
-			).put(
-				"[%BILLING_ADDRESS_NAME%]", billingAddress.getName()
-			).put(
-				"[%BILLING_ADDRESS_PHONE%]", billingAddress.getPhoneNumber()
-			).put(
-				"[%CATALOG_NAME%]", catalog.getName()
-			).put(
-				"[%EMAIL_ADDRESS%]", order.getCreatorEmailAddress()
-			).put(
-				"[%EXCHANGE_RATE%]", _getExchangeRate(order)
-			).put(
-				"[%LICENSE_TYPE%]", productSpecificationsMap.get("license-type")
-			).put(
-				"[%NET_PRICE_FORMATTED%]", order.getSubtotalFormatted()
-			).put(
-				"[%ORDER_DATE%]", _format(order.getCreateDate())
-			).put(
-				"[%ORDER_ID%]", String.valueOf(order.getId())
-			).put(
-				"[%ORDER_PAYMENT_METHOD%]",
-				MarketplaceConstants.getOrderPaymentMethodLabel(
-					order.getPaymentMethod())
-			).put(
-				"[%ORDER_STATUS%]",
-				MarketplaceConstants.getOrderStatusLabel(order.getOrderStatus())
-			).put(
-				"[%PAYMENT_TERMS%]", order.getPaymentTermDescription()
-			).put(
-				"[%PRODUCT_THUMBNAIL%]",
-				new URL(
-					StringBundler.concat(
-						lxcDXPServerProtocol, "://", lxcDXPMainDomain,
-						product.getThumbnail())
-				).toString(
-				).replaceAll(
-					"(?<=accounts/)-?\\d+(?=/images)", "-1"
-				)
-			).put(
-				"[%SUBSCRIPTION_EXPIRATION_DATE%]",
-				_format(
-					MarketplaceUtil.getOrderPurchaseEndDate(
-						productSpecificationsMap.get("license-type"),
-						MarketplaceUtil.getSkuOptionValue(
-							"license-usage-type", orderItem.getOptions())))
-			).put(
-				"[%SUBSCRIPTION_STARTING_DATE%]", _format(order.getCreateDate())
-			).put(
-				"[%SUBSCRIPTION_TYPE%]",
-				productSpecificationsMap.get("license-type")
-			).put(
-				"[%TOTAL_FORMATTED%]", order.getTotalFormatted()
-			).put(
-				"[%VAT_FORMATTED%]", order.getTaxAmountFormatted()
-			).put(
-				"[%VAT_NUMBER%]", account.getTaxId()
-			).build());
-	}
-
-	private void _sendPaymentApprovedNotification(
-			Order order, String emailDescription, String buttonText,
-			String json)
-		throws Exception {
-
-		OrderItem[] orderItems = order.getOrderItems();
-
-		OrderItem orderItem = orderItems[0];
-
-		if (orderItem == null) {
-			return;
-		}
-
-		if (_log.isInfoEnabled()) {
-			_log.info(
-				"Sending payment approved notification for order " +
-					order.getId());
-		}
-
-		Product product = _marketplaceService.getProductBySkuId(
-			orderItem.getSkuId());
-
-		Catalog catalog = _marketplaceService.getCatalog(
-			product.getCatalogId());
-
-		JSONObject jsonObject = new JSONObject(json);
-
-		JSONObject modelCPDefinitionJSONObject = jsonObject.getJSONObject(
-			"modelCPDefinition");
-
-		_marketplaceService.postNotificationQueueEntry(
-			null, "MARKETPLACE-PAYMENT-APPROVED",
-			new HashMapBuilder<String, String>().put(
-				"[%APP_NAME%]",
-				product.getName(
-				).get(
-					"en_US"
-				)
-			).put(
-				"[%CATALOG_NAME%]", catalog.getName()
-			).put(
-				"[%COMMERCEORDER_AUTHOR_EMAIL_ADDRESS%]",
-				order.getCreatorEmailAddress()
-			).put(
-				"[%APP_NET_PRICE%]", order.getSubtotalFormatted()
-			).put(
-				"[%ORDER_ID%]", String.valueOf(order.getId())
-			).put(
-				"[%PRODUCT_THUMBNAIL%]",
-				new URL(
-					StringBundler.concat(
-						lxcDXPServerProtocol, "://", lxcDXPMainDomain,
-						product.getThumbnail())
-				).toString(
-				).replaceAll(
-					"(?<=accounts/)-?\\d+(?=/images)", "-1"
-				)
-			).put(
-				"[%EMAIL_DESCRIPTION%]", emailDescription
-			).put(
-				"[%BUTTON_TEXT%]", buttonText
-			).put(
-				"[%APP_TOTAL_PRICE%]", order.getTotalFormatted()
-			).put(
-				"[%APP_VAT%]", order.getTaxAmountFormatted()
-			).put(
-				"[%CPDEFINITION_ID%]",
-				String.valueOf(
-					modelCPDefinitionJSONObject.getLong("CPDefinitionId"))
-			).build());
-	}
-
 	private void _setExchangeRate(Order order) throws Exception {
 		Map<String, String> customFields =
 			(Map<String, String>)order.getCustomFields();
@@ -1050,134 +593,6 @@ public class MarketplaceRestController extends BaseRestController {
 			).toString());
 	}
 
-	private void _setUpAddOns(
-			Jwt jwt, Order order, Map<String, String> productSpecificationsMap)
-		throws Exception {
-
-		String solutionType = productSpecificationsMap.get("solution-type");
-
-		if (Objects.equals(solutionType, "analytics")) {
-			_setUpAnalyticsAddOn(jwt, order);
-
-			return;
-		}
-
-		if (Objects.equals(solutionType, "ai-hub") ||
-			Objects.equals(solutionType, "content-data-platform")) {
-
-			_setUpCustomAddOn(
-				productSpecificationsMap.get("license-type"), order);
-		}
-	}
-
-	private void _setUpAnalyticsAddOn(Jwt jwt, Order order) throws Exception {
-		if (!order.getAccountExternalReferenceCode(
-			).startsWith(
-				"KOR-"
-			)) {
-
-			return;
-		}
-
-		Map<String, String> customFields =
-			(Map<String, String>)order.getCustomFields();
-
-		JSONObject orderMetadataJSONObject = new JSONObject(
-			customFields.getOrDefault("order-metadata", "{}"));
-
-		if (_koroneikiService.hasEntitlement(
-				_koroneikiService.getKoroneikiAccount(
-					order.getAccountExternalReferenceCode()),
-				MarketplaceConstants.KORONEIKI_AC_ENTITLEMENTS)) {
-
-			_koroneikiService.linkProductPurchaseToOpportunity(
-				jwt, String.valueOf(order.getId()),
-				orderMetadataJSONObject.getString("productPurchaseKey"));
-
-			return;
-		}
-
-		for (OrderItem orderItem : order.getOrderItems()) {
-			if (!Objects.equals(
-					orderItem.getSkuExternalReferenceCode(),
-					orderMetadataJSONObject.getString("productKey"))) {
-
-				continue;
-			}
-
-			_koroneikiService.postAccountAccountKeyProductPurchase(
-				order.getAccountExternalReferenceCode(), jwt, "Subscription",
-				MarketplaceUtil.getSkuOptionValue(
-					"license-usage-type", orderItem.getOptions()),
-				orderItem);
-		}
-	}
-
-	private void _setUpCustomAddOn(String licenseType, Order order)
-		throws Exception {
-
-		OrderItem[] orderItems = order.getOrderItems();
-
-		OrderItem orderItem = orderItems[0];
-
-		if (orderItem == null) {
-			return;
-		}
-
-		UserAccount userAccount = _marketplaceService.getUserAccount(
-			order.getCreatorEmailAddress());
-
-		Product product = _marketplaceService.getProductBySkuId(
-			orderItem.getSkuId());
-
-		_salesforceService.postSalesforceOpportunity(
-			new SalesforceOpportunity(
-				licenseType, order, orderItem, product, userAccount));
-	}
-
-	private void _setUpProductEntitlements(
-			Jwt jwt, String licenseType, Order order)
-		throws Exception {
-
-		String accountExternalReferenceCode =
-			order.getAccountExternalReferenceCode();
-
-		if (!accountExternalReferenceCode.startsWith("KOR-")) {
-			AccountResource accountResource =
-				_marketplaceService.getAccountResource();
-
-			Account account = accountResource.getAccount(order.getAccountId());
-
-			accountResource.patchAccount(
-				account.getId(),
-				new Account() {
-					{
-						setExternalReferenceCode(
-							() -> _koroneikiService.postKoroneikiAccount(
-								account, jwt
-							).getKey());
-					}
-				});
-		}
-
-		try {
-			for (OrderItem orderItem : order.getOrderItems()) {
-				_koroneikiService.postAccountAccountKeyProductPurchase(
-					accountExternalReferenceCode, jwt, licenseType,
-					MarketplaceUtil.getSkuOptionValue(
-						"license-usage-type", orderItem.getOptions()),
-					orderItem);
-			}
-
-			_marketplaceService.updateOrder(
-				null, order.getId(),
-				MarketplaceConstants.ORDER_STATUS_COMPLETED);
-		}
-		catch (Exception exception) {
-			_log.error("Unable to create account product purchase", exception);
-		}
-	}
-
 	private static final int _ACCOUNT_TYPE_BUSINESS = 2;
 
 	private static final int _ACCOUNT_TYPE_PERSON = 1;
@@ -1197,8 +612,5 @@ public class MarketplaceRestController extends BaseRestController {
 
 	@Autowired
 	private MarketplaceService _marketplaceService;
-
-	@Autowired
-	private SalesforceService _salesforceService;
 
 }
